backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
supabase==2.18.1
python-multipart==0.0.20
pydantic[email]==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
sqlalchemy==2.0.43 